backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
openai
google-generativeai